use the native thread pool for crt compilation and pipelining
Moved ThreadPool to near-async (core/async/src/thread_pool.rs)
  - Identical to the original soft_realtime_thread_pool implementation — OS threads, SCHED_RR, Mutex+Condvar+VecDeque work queue, idle timeout, metrics
  - Made pub so it can be used from other crates
  - thread-priority dependency moved from chain/chain to core/async

  Slimmed chain/chain/src/soft_realtime_thread_pool.rs
  - Now just thin wrappers (ApplyChunksSpawner, PartialWitnessValidationThreadPool, WitnessCreationThreadPool) importing ThreadPool from near_async

  Shared pool for compilation + pipelining
  - NightshadeRuntime owns Arc<ThreadPool> named "contract_preparation"
  - precompile_contracts() spawns jobs on it with crossbeam::sync::WaitGroup to wait for completion
  - ApplyState carries Option<Arc<dyn AsyncComputationSpawner>> to pass the pool to pipelining
  - ReceiptPreparationPipeline uses pool.spawn_boxed("crt-tasks", ...) instead of rayon::spawn_fifo
  - When pool is None (tests, view calls), tasks stay Pending and get_contract() prepares on the main thread

s# Please enter the commit message for your changes. Lines starting
