+++ chain/chain/Cargo.toml
@@ -27,7 +27,6 @@ serde.workspace = true
 strum.workspace = true
 tempfile.workspace = true
 thiserror.workspace = true
-thread-priority.workspace = true
 time.workspace = true
 tokio.workspace = true
 tracing.workspace = true
