+++ integration-tests/src/user/runtime_user.rs
@@ -208,6 +208,7 @@ impl RuntimeUser {
             bandwidth_requests: BlockBandwidthRequests::empty(),
             trie_access_tracker_state: Default::default(),
             on_post_state_ready: None,
+            contract_preparation_pool: None,
         }
     }
