+++ runtime/runtime/src/access_keys.rs
@@ -396,6 +396,7 @@ mod tests {
             bandwidth_requests: BlockBandwidthRequests::empty(),
             trie_access_tracker_state: Default::default(),
             on_post_state_ready: None,
+            contract_preparation_pool: None,
         }
     }
