+++ runtime/runtime/src/actions.rs
@@ -1078,6 +1078,7 @@ mod tests {
             bandwidth_requests: BlockBandwidthRequests::empty(),
             trie_access_tracker_state: Default::default(),
             on_post_state_ready: None,
+            contract_preparation_pool: None,
         }
     }
