+++ runtime/runtime/src/pipelining.rs
@@ -6,6 +6,7 @@ use crate::metrics::{
     PIPELINING_ACTIONS_SUBMITTED, PIPELINING_ACTIONS_TASK_DELAY_TIME,
     PIPELINING_ACTIONS_TASK_WORKING_TIME, PIPELINING_ACTIONS_WAITING_TIME,
 };
+use near_async::thread_pool::ThreadPool;
 use near_parameters::RuntimeConfig;
 use near_primitives::account::{Account, AccountContract};
 use near_primitives::action::{Action, GlobalContractIdentifier};
@@ -21,21 +22,8 @@ use near_vm_runner::logic::GasCounter;
 use near_vm_runner::{ContractRuntimeCache, PreparedContract};
 use parking_lot::{Condvar, Mutex};
 use std::collections::{BTreeMap, BTreeSet, HashSet};
-use std::sync::{Arc, OnceLock};
+use std::sync::Arc;
 use std::time::Instant;
-
-/// Dedicated thread pool for pipelining, to avoid sharing the global rayon pool
-/// with wasmtime's internal compilation which also uses rayon.
-fn pipelining_pool() -> &'static rayon::ThreadPool {
-    static POOL: OnceLock<rayon::ThreadPool> = OnceLock::new();
-    POOL.get_or_init(|| {
-        rayon::ThreadPoolBuilder::new()
-            .num_threads(std::cmp::max(rayon::current_num_threads() / 2, 1))
-            .thread_name(|index| format!("pipelining-{index}"))
-            .build()
-            .expect("failed to build pipelining thread pool")
-    })
-}
 
 pub(crate) struct ReceiptPreparationPipeline {
     /// Mapping from a Receipt's ID to a parallel "task" to prepare the receipt's data.
@@ -81,6 +69,10 @@ pub(crate) struct ReceiptPreparationPipeline {
     storage: ContractStorage,
 
     chain_id: String,
+
+    /// Thread pool for contract preparation. When `None`, tasks stay `Pending`
+    /// and `get_contract` will prepare on the main thread.
+    pool: Option<Arc<ThreadPool>>,
 }
 
 #[derive(PartialEq, Eq, PartialOrd, Ord)]
@@ -107,6 +99,7 @@ impl ReceiptPreparationPipeline {
         contract_cache: Option<Box<dyn ContractRuntimeCache>>,
         storage: ContractStorage,
         chain_id: String,
+        pool: Option<Arc<ThreadPool>>,
     ) -> Self {
         Self {
             map: Default::default(),
@@ -116,6 +109,7 @@ impl ReceiptPreparationPipeline {
             contract_cache,
             storage,
             chain_id,
+            pool,
         }
     }
 
@@ -225,30 +219,33 @@ impl ReceiptPreparationPipeline {
                     let task = Arc::new(PrepareTask { status, condvar: Condvar::new() });
                     entry.insert(Arc::clone(&task));
                     PIPELINING_ACTIONS_SUBMITTED.inc_by(1);
-                    pipelining_pool().spawn_fifo(move || {
-                        let task_status = {
+                    if let Some(pool) = &self.pool {
+                        pool.spawn_boxed(Box::new(move || {
+                            let task_status = {
+                                let mut status = task.status.lock();
+                                std::mem::replace(&mut *status, PrepareTaskStatus::Working)
+                            };
+                            let PrepareTaskStatus::Pending = task_status else {
+                                return;
+                            };
+                            PIPELINING_ACTIONS_TASK_DELAY_TIME
+                                .inc_by(created.elapsed().as_secs_f64());
+                            let start = Instant::now();
+                            let contract = prepare_function_call(
+                                &storage,
+                                cache.as_deref(),
+                                config,
+                                gas_counter,
+                                identifier,
+                                &method_name,
+                            );
                             let mut status = task.status.lock();
-                            std::mem::replace(&mut *status, PrepareTaskStatus::Working)
-                        };
-                        let PrepareTaskStatus::Pending = task_status else {
-                            return;
-                        };
-                        PIPELINING_ACTIONS_TASK_DELAY_TIME.inc_by(created.elapsed().as_secs_f64());
-                        let start = Instant::now();
-                        let contract = prepare_function_call(
-                            &storage,
-                            cache.as_deref(),
-                            config,
-                            gas_counter,
-                            identifier,
-                            &method_name,
-                        );
-
-                        let mut status = task.status.lock();
-                        *status = PrepareTaskStatus::Prepared(contract);
-                        PIPELINING_ACTIONS_TASK_WORKING_TIME.inc_by(start.elapsed().as_secs_f64());
-                        task.condvar.notify_all();
-                    });
+                            *status = PrepareTaskStatus::Prepared(contract);
+                            PIPELINING_ACTIONS_TASK_WORKING_TIME
+                                .inc_by(start.elapsed().as_secs_f64());
+                            task.condvar.notify_all();
+                        }));
+                    }
                     any_function_calls = true;
                 }
                 // No need to handle this receipt as it only generates other new receipts.
