+++ runtime/runtime/src/verifier.rs
@@ -1130,6 +1130,7 @@ mod tests {
             bandwidth_requests: BlockBandwidthRequests::empty(),
             trie_access_tracker_state: Default::default(),
             on_post_state_ready: None,
+            contract_preparation_pool: None,
         }
     }
