+++ runtime/runtime/tests/runtime_group_tools/mod.rs
@@ -118,6 +118,7 @@ impl StandaloneRuntime {
             bandwidth_requests: BlockBandwidthRequests::empty(),
             trie_access_tracker_state: Default::default(),
             on_post_state_ready: None,
+            contract_preparation_pool: None,
         };
 
         Self {
