+++ test-utils/testlib/src/state_witness_test_data.rs
@@ -94,6 +94,7 @@ pub fn generate_realistic_state_witness(target_size_bytes: usize) -> ChunkStateW
         bandwidth_requests: BlockBandwidthRequests::empty(),
         trie_access_tracker_state: Default::default(),
         on_post_state_ready: None,
+        contract_preparation_pool: None,
     };
 
     // Collect data for building the witness
